refactor(commands): migrate attach/detach/setup/kill, daemon lifecycle, update
Issue #282 steps 4–6, batched because updateCmd depends on daemonRestart
(the lifecycle helpers landed together with their consumers):

- attach/detach: AttachCmd/DetachCmd + hook-strip logic (#168, deprecated
  entry point #214 preserved verbatim)
- setup: SetupCmd + windows task path, with systemctlFn/osExecutableFn seams
- kill: KillCmd
- daemon lifecycle: DaemonCmd, daemonStart/Stop/Restart, EnsureDaemon/
  WithDaemon (file-locked lazy start), forceKillDaemon, and the lock/kill/
  procattr platform files; RestartDaemonAfterLogin extracted from the auth
  flow's inline restart helper
- update: UpdateCmd + download/checksum/replace helpers with the
  updateAPIBase/updateDownloadBase seams

All tests move with their commands (exported entry points, same assertions).
cmd/riffpad keeps one-line wrappers the dispatch still references.

Pure relocation; no behavior change.
