+++ apps/daemon/cmd/riffpad/main.go
@@ -5,8 +5,6 @@ package main
 import (
 	"bytes"
 	"context"
-	"crypto/sha256"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"flag"
@@ -20,7 +18,6 @@ import (
 	"os/signal"
 	"path/filepath"
 	"runtime"
-	"strconv"
 	"strings"
 	"syscall"
 	"time"
@@ -44,16 +41,6 @@ const updateRepo = "riffpad/riffpad"
 var (
 	updateAPIBase      = "https://api.github.com/repos"
 	updateDownloadBase = "https://github.com/" + updateRepo + "/releases/latest/download/"
-)
-
-// osExecutableFn and systemctlFn are indirections so update/setup tests can
-// run against a throwaway binary and without touching the real systemd user
-// session.
-var (
-	osExecutableFn = os.Executable
-	systemctlFn    = func(args ...string) ([]byte, error) {
-		return exec.Command("systemctl", append([]string{"--user"}, args...)...).CombinedOutput()
-	}
 )
 
 // pairRetryDelay and pairRetryMaxWait control the transient "host offline"
@@ -282,339 +269,11 @@ func extractLangFlag(args []string) (lang string, rest []string) {
 }
 
 func daemonCmd(args []string, base, dataDir string) error {
-	if len(args) < 1 {
-		return fmt.Errorf("%s", t.T("usage_daemon"))
-	}
-	switch args[0] {
-	case "start":
-		return daemonStart(base, dataDir)
-	case "stop":
-		return daemonStop(base, dataDir)
-	case "restart":
-		return daemonRestart(base, dataDir)
-	default:
-		return fmt.Errorf("%s", t.T("usage_daemon"))
-	}
+	return commands.DaemonCmd(args, base, dataDir)
 }
 
-// systemdActiveFn reports whether the riffpad systemd user service is active.
-// Indirection lets tests simulate both managed and unmanaged daemons.
-var systemdActiveFn = func() (bool, error) {
-	out, err := exec.Command("systemctl", "--user", "is-active", "riffpad.service").CombinedOutput()
-	if err != nil {
-		return false, err
-	}
-	return strings.TrimSpace(string(out)) == "active", nil
-}
-
-// systemdRestartFn restarts the riffpad systemd user service.
-var systemdRestartFn = func() error {
-	return exec.Command("systemctl", "--user", "restart", "riffpad.service").Run()
-}
-
-// windowsTaskExistsFn reports whether the RiffpadDaemon scheduled task
-// (created by `riffpad setup` / install.ps1) exists.
-var windowsTaskExistsFn = func() (bool, error) {
-	if err := exec.Command("schtasks", "/Query", "/TN", "RiffpadDaemon").Run(); err != nil {
-		return false, err
-	}
-	return true, nil
-}
-
-// windowsTaskRunFn starts the RiffpadDaemon scheduled task so Task Scheduler
-// stays the bookkeeper for the daemon process.
-var windowsTaskRunFn = func() error {
-	return exec.Command("schtasks", "/Run", "/TN", "RiffpadDaemon").Run()
-}
-
-// daemonRestart restarts the daemon. When the systemd user service is active,
-// it goes through systemctl so systemd stays the single source of truth;
-// on Windows, an existing RiffpadDaemon scheduled task is restarted through
-// Task Scheduler; otherwise it stops and starts the background process. A
-// daemon that is not running is an error.
-func daemonRestart(base, dataDir string) error {
-	switch runtime.GOOS {
-	case "linux":
-		if active, err := systemdActiveFn(); err == nil && active {
-			return restartViaSystemd()
-		}
-	case "windows":
-		if ok, err := windowsTaskExistsFn(); err == nil && ok {
-			return restartViaWindowsTask(base, dataDir)
-		}
-	}
-	if !reachable(base) {
-		return fmt.Errorf("%s", t.T("daemon_not_running"))
-	}
-	if err := daemonStop(base, dataDir); err != nil {
-		return fmt.Errorf("%s: %w", t.T("daemon_restart_failed"), err)
-	}
-	if err := daemonStart(base, dataDir); err != nil {
-		return fmt.Errorf("%s: %w", t.T("daemon_restart_failed"), err)
-	}
-	fmt.Println(t.T("daemon_restarted"))
-	return nil
-}
-
-func restartViaSystemd() error {
-	if err := systemdRestartFn(); err != nil {
-		return fmt.Errorf("%s: %w", t.T("daemon_restart_failed"), err)
-	}
-	fmt.Println(t.T("daemon_restarted"))
-	return nil
-}
-
-func restartViaWindowsTask(base, dataDir string) error {
-	// Stop the current daemon first (if any) so the task-started process can
-	// bind the port; Task Scheduler then owns the new instance.
-	if reachable(base) {
-		if err := daemonStop(base, dataDir); err != nil {
-			return fmt.Errorf("%s: %w", t.T("daemon_restart_failed"), err)
-		}
-	}
-	if err := windowsTaskRunFn(); err != nil {
-		return fmt.Errorf("%s: %w", t.T("daemon_restart_failed"), err)
-	}
-	if !waitReachable(base, 2*time.Second) {
-		return fmt.Errorf("%s", t.T("daemon_restart_wait_failed"))
-	}
-	fmt.Println(t.T("daemon_restarted"))
-	return nil
-}
-
-func waitReachable(base string, timeout time.Duration) bool {
-	deadline := time.Now().Add(timeout)
-	for time.Now().Before(deadline) {
-		time.Sleep(100 * time.Millisecond)
-		if reachable(base) {
-			return true
-		}
-	}
-	return false
-}
-
-func daemonStart(base, dataDir string) error {
-	if reachable(base) {
-		return fmt.Errorf("%s", t.T("daemon_already_running", base))
-	}
-	exe, err := os.Executable()
-	if err != nil {
-		return err
-	}
-	logDir := filepath.Join(dataDir, "logs")
-	if err := os.MkdirAll(logDir, 0o700); err != nil {
-		return err
-	}
-	out, err := os.OpenFile(filepath.Join(logDir, "daemon.out.log"), os.O_CREATE|os.O_APPEND|os.O_WRONLY, 0o600)
-	if err != nil {
-		return err
-	}
-	defer out.Close()
-	cmd := exec.Command(exe, "_daemon", "--data-dir", dataDir)
-	cmd.Stdout = out
-	cmd.Stderr = out
-	cmd.SysProcAttr = daemonProcAttr()
-	if err := cmd.Start(); err != nil {
-		return fmt.Errorf("start daemon: %w", err)
-	}
-	_ = os.WriteFile(filepath.Join(dataDir, "daemon.pid"), []byte(fmt.Sprintf("%d\n", cmd.Process.Pid)), 0o600)
-	for i := 0; i < 20; i++ {
-		time.Sleep(100 * time.Millisecond)
-		if reachable(base) {
-			fmt.Println(t.T("daemon_started", base))
-			return nil
-		}
-	}
-	return fmt.Errorf("%s", t.T("daemon_start_wait_failed", filepath.Join(logDir, "daemon.out.log")))
-}
-
-// startDaemonFn is indirection so tests can observe/emulate lazy starts.
-var startDaemonFn = daemonStart
-
-// withDaemon ensures the daemon is running before executing an operation that
-// needs it. Only reachable checks and daemon start are guarded by a file lock,
-// so concurrent riffpad invocations start at most one daemon.
 func withDaemon(fn func() error, base, dataDir string) error {
-	if err := ensureDaemon(base, dataDir); err != nil {
-		return err
-	}
-	return fn()
-}
-
-func ensureDaemon(base, dataDir string) error {
-	if reachable(base) {
-		return nil
-	}
-	if err := os.MkdirAll(dataDir, 0o700); err != nil {
-		return err
-	}
-	lockPath := filepath.Join(dataDir, "daemon.lock")
-	lf, err := os.OpenFile(lockPath, os.O_CREATE|os.O_RDWR, 0o600)
-	if err != nil {
-		return fmt.Errorf("open daemon lock: %w", err)
-	}
-	defer lf.Close()
-	if err := lockFile(lf); err != nil {
-		return fmt.Errorf("lock daemon start: %w", err)
-	}
-	defer func() { _ = unlockFile(lf) }()
-	// Another process may have started the daemon while we waited for the lock.
-	if reachable(base) {
-		return nil
-	}
-	if err := startDaemonFn(base, dataDir); err != nil {
-		return err
-	}
-	return nil
-}
-
-// setupCmd installs (or removes) a Linux systemd user service so the daemon
-// starts at login and restarts after crashes.
-func setupCmd(args []string, dataDir string) error {
-	fs := flag.NewFlagSet("setup", flag.ExitOnError)
-	remove := fs.Bool("remove", false, "stop and remove the autostart entry")
-	_ = fs.Parse(args)
-	if runtime.GOOS == "windows" {
-		return setupWindowsTask(*remove, dataDir)
-	}
-	if runtime.GOOS != "linux" {
-		return fmt.Errorf("setup currently supports Linux systemd / Windows Task Scheduler only")
-	}
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return err
-	}
-	unitDir := filepath.Join(home, ".config", "systemd", "user")
-	unitPath := filepath.Join(unitDir, "riffpad.service")
-	if *remove {
-		_, _ = systemctlFn("disable", "--now", "riffpad.service")
-		if err := os.Remove(unitPath); err != nil && !os.IsNotExist(err) {
-			return err
-		}
-		fmt.Println(t.T("setup_removed"))
-		return nil
-	}
-	exe, err := osExecutableFn()
-	if err != nil {
-		return err
-	}
-	if err := os.MkdirAll(unitDir, 0o700); err != nil {
-		return err
-	}
-	execStart := exe
-	if strings.Contains(exe, " ") {
-		execStart = `"` + exe + `"`
-	}
-	// The systemd user manager runs with a minimal PATH, which would hide
-	// coding CLIs installed in user dirs (~/.local/bin etc.). Capture the
-	// PATH of the shell that ran `riffpad setup` so the daemon can spawn
-	// codex/kimi/claude. "%" is a systemd specifier, so escape it.
-	path := strings.ReplaceAll(os.Getenv("PATH"), "%", "%%")
-	unit := fmt.Sprintf(`[Unit]
-Description=Riffpad daemon (AI agent remote control)
-After=network-online.target
-
-[Service]
-Environment=PATH=%s
-ExecStart=%s _daemon --data-dir %s
-Restart=on-failure
-RestartSec=2
-
-[Install]
-WantedBy=default.target
-`, path, execStart, dataDir)
-	if err := os.WriteFile(unitPath, []byte(unit), 0o600); err != nil {
-		return err
-	}
-	if out, err := systemctlFn("daemon-reload"); err != nil {
-		return fmt.Errorf("systemctl daemon-reload: %v\n%s", err, out)
-	}
-	if out, err := systemctlFn("enable", "--now", "riffpad.service"); err != nil {
-		return fmt.Errorf("systemctl enable riffpad: %v\n%s", err, out)
-	}
-	fmt.Println(t.T("setup_installed", unitPath))
-	fmt.Println(t.T("setup_done"))
-	return nil
-}
-
-// setupWindowsTask registers (or removes) a scheduled task that starts the
-// daemon at logon.
-func setupWindowsTask(remove bool, dataDir string) error {
-	if remove {
-		out, err := exec.Command("schtasks", "/Delete", "/TN", "RiffpadDaemon", "/F").CombinedOutput()
-		if err != nil {
-			return fmt.Errorf("schtasks delete: %v\n%s", err, out)
-		}
-		fmt.Println(t.T("setup_removed"))
-		return nil
-	}
-	exe, err := os.Executable()
-	if err != nil {
-		return err
-	}
-	tr := fmt.Sprintf(`"%s" _daemon --data-dir "%s"`, exe, dataDir)
-	out, err := exec.Command("schtasks", "/Create", "/TN", "RiffpadDaemon", "/TR", tr, "/SC", "ONLOGON", "/RL", "LIMITED", "/F").CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("schtasks create: %v\n%s", err, out)
-	}
-	fmt.Println(t.T("setup_installed", "RiffpadDaemon"))
-	fmt.Println(t.T("setup_done"))
-	return nil
-}
-
-func daemonStop(base, dataDir string) error {
-	if !reachable(base) {
-		return fmt.Errorf("%s", t.T("daemon_not_running"))
-	}
-	// Bound the shutdown request itself: a wedged daemon may accept the
-	// connection yet never answer, and http.DefaultClient has no timeout.
-	client := &http.Client{Timeout: 5 * time.Second}
-	resp, err := daemonDo(client, http.MethodPost, base+"/api/shutdown", nil)
-	if err == nil {
-		io.Copy(io.Discard, resp.Body)
-		resp.Body.Close()
-	}
-	for i := 0; i < 20; i++ {
-		time.Sleep(100 * time.Millisecond)
-		if !reachable(base) {
-			fmt.Println(t.T("daemon_stopped"))
-			return nil
-		}
-	}
-	// HTTP shutdown failed or the daemon ignored it; fall back to the pid
-	// file written by daemonStart and kill the process (#174).
-	if err := forceKillDaemon(dataDir); err != nil {
-		return err
-	}
-	for i := 0; i < 20; i++ {
-		time.Sleep(100 * time.Millisecond)
-		if !reachable(base) {
-			fmt.Println(t.T("daemon_stopped"))
-			return nil
-		}
-	}
-	return fmt.Errorf("%s", t.T("daemon_did_not_stop"))
-}
-
-// forceKillDaemon reads the pid file written by daemonStart and terminates
-// the process after verifying it really is a riffpad daemon, so a stale or
-// recycled pid can never take down an unrelated process (#174).
-func forceKillDaemon(dataDir string) error {
-	raw, err := os.ReadFile(filepath.Join(dataDir, "daemon.pid"))
-	if err != nil {
-		return fmt.Errorf("%s", t.T("daemon_stop_no_pid"))
-	}
-	pid, err := strconv.Atoi(strings.TrimSpace(string(raw)))
-	if err != nil || pid <= 0 {
-		return fmt.Errorf("%s", t.T("daemon_stop_no_pid"))
-	}
-	if !isRiffpadDaemonProcess(pid) {
-		return fmt.Errorf("%s", t.T("daemon_stop_pid_mismatch", pid))
-	}
-	if err := terminateProcess(pid); err != nil {
-		return fmt.Errorf("%s", t.T("daemon_stop_kill_failed", err))
-	}
-	return nil
+	return commands.WithDaemon(fn, base, dataDir)
 }
 
 func statusCmd(base string) error { return commands.StatusCmd(base) }
@@ -627,195 +286,11 @@ func runCmd(args []string, base string) error { return commands.RunCmd(args, bas
 
 func logsCmd(dataDir string) error { return commands.LogsCmd(dataDir) }
 
-// attachCmd injects Claude Code hooks pointing at the local daemon, so a
-// normal interactive `claude` session is captured and approvals can be made
-// from the web UI / mobile. Existing user hooks are preserved: only riffpad's
-// own entries are replaced, making repeated attaches idempotent.
-func attachCmd(base string) error {
-	// Attach mode is deprecated: it permanently edits the user's global
-	// Claude settings, which breaks normal claude startup when the daemon is
-	// off or unpaired. The implementation is kept below for reference; only
-	// the entry point is disabled (#214).
-	return fmt.Errorf("%s", t.T("attach_deprecated"))
-	if !reachable(base) {
-		return fmt.Errorf("%s", t.T("daemon_start_hint", base))
-	}
-	port := defaultDaemonPort(base)
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return err
-	}
-	settingsPath := filepath.Join(home, ".claude", "settings.json")
-	if err := os.MkdirAll(filepath.Dir(settingsPath), 0o700); err != nil {
-		return err
-	}
-	settings := map[string]any{}
-	if data, err := os.ReadFile(settingsPath); err == nil {
-		if err := json.Unmarshal(data, &settings); err != nil {
-			return fmt.Errorf("parse %s: %w", settingsPath, err)
-		}
-	}
-	backup := settingsPath + ".riffpad.bak"
-	if _, err := os.Stat(backup); os.IsNotExist(err) {
-		raw, _ := json.MarshalIndent(settings, "", "  ")
-		if err := os.WriteFile(backup, raw, 0o600); err != nil {
-			return err
-		}
-	}
-	baseURL := fmt.Sprintf("http://127.0.0.1:%d", port)
-	httpHook := func(path string, timeout int) map[string]any {
-		hookURL := baseURL + path
-		if tok := localToken(); tok != "" {
-			// The hook process cannot set headers; the daemon also accepts the
-			// local API token as a query parameter.
-			hookURL += "?token=" + url.QueryEscape(tok)
-		}
-		return map[string]any{
-			"matcher": "",
-			"hooks": []any{
-				map[string]any{"type": "http", "url": hookURL, "timeout": timeout},
-			},
-		}
-	}
-	// Merge, don't overwrite: drop any previously injected riffpad entries
-	// (idempotent re-attach) while keeping the user's own hooks intact.
-	if stripRiffpadHooks(settings) {
-		fmt.Println(t.T("attach_keep_existing"))
-	}
-	hooks, ok := settings["hooks"].(map[string]any)
-	if !ok {
-		hooks = map[string]any{}
-		settings["hooks"] = hooks
-	}
-	add := func(event string, entry map[string]any) {
-		list, _ := hooks[event].([]any)
-		hooks[event] = append(list, entry)
-	}
-	add("SessionStart", httpHook("/hooks/claude/session-start", 10))
-	add("SessionEnd", httpHook("/hooks/claude/session-end", 10))
-	add("UserPromptSubmit", httpHook("/hooks/claude/user-prompt-submit", 30))
-	add("MessageDisplay", httpHook("/hooks/claude/message-display", 10))
-	add("PreToolUse", httpHook("/hooks/claude/pre-tool-use", 10))
-	add("PostToolUse", httpHook("/hooks/claude/post-tool-use", 10))
-	add("PermissionRequest", httpHook("/hooks/claude/permission", 600))
-	add("Notification", httpHook("/hooks/claude/notification", 10))
-	raw, err := json.MarshalIndent(settings, "", "  ")
-	if err != nil {
-		return err
-	}
-	if err := os.WriteFile(settingsPath, raw, 0o600); err != nil {
-		return err
-	}
-	fmt.Println(t.T("attach_injected"))
-	fmt.Println(t.T("attach_next"))
-	fmt.Println(t.T("attach_verify"))
-	return nil
-}
+func attachCmd(base string) error { return commands.AttachCmd(base) }
 
-// riffpadHookPathPrefix namespaces every hook URL the daemon injects, so
-// attach/detach can tell riffpad's own entries apart from the user's — query
-// parameters (e.g. the ?token= added for local auth) are ignored by matching
-// on the parsed URL path only.
-const riffpadHookPathPrefix = "/hooks/claude/"
+func detachCmd() error { return commands.DetachCmd() }
 
-func isRiffpadHook(rawURL string) bool {
-	u, err := url.Parse(rawURL)
-	return err == nil && strings.HasPrefix(u.Path, riffpadHookPathPrefix)
-}
-
-// stripRiffpadHooks removes riffpad-injected hook entries from
-// settings["hooks"] in place, keeping everything the user added. It reports
-// whether any user-defined (non-riffpad) hooks remain.
-func stripRiffpadHooks(settings map[string]any) bool {
-	hooks, ok := settings["hooks"].(map[string]any)
-	if !ok {
-		return false
-	}
-	userHooks := false
-	for event, list := range hooks {
-		entries, ok := list.([]any)
-		if !ok {
-			userHooks = true
-			continue
-		}
-		kept := make([]any, 0, len(entries))
-		for _, e := range entries {
-			entry, ok := e.(map[string]any)
-			if !ok {
-				kept = append(kept, e)
-				continue
-			}
-			nested, ok := entry["hooks"].([]any)
-			if !ok {
-				kept = append(kept, e)
-				continue
-			}
-			keptNested := make([]any, 0, len(nested))
-			for _, h := range nested {
-				if hm, ok := h.(map[string]any); ok {
-					if raw, _ := hm["url"].(string); raw != "" && isRiffpadHook(raw) {
-						continue
-					}
-				}
-				keptNested = append(keptNested, h)
-			}
-			if len(keptNested) == 0 {
-				continue
-			}
-			entry["hooks"] = keptNested
-			kept = append(kept, entry)
-		}
-		if len(kept) == 0 {
-			delete(hooks, event)
-			continue
-		}
-		hooks[event] = kept
-		userHooks = true
-	}
-	if len(hooks) == 0 {
-		delete(settings, "hooks")
-	}
-	return userHooks
-}
-
-// detachCmd removes only the riffpad-injected hook entries, leaving any user
-// configuration (including hooks added after attach) untouched. The
-// settings.json.riffpad.bak snapshot from the first attach is no longer used
-// for restoration; it is kept on disk for manual reference.
-func detachCmd() error {
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return err
-	}
-	settingsPath := filepath.Join(home, ".claude", "settings.json")
-	data, err := os.ReadFile(settingsPath)
-	if err != nil {
-		return fmt.Errorf("no settings found (nothing to detach)")
-	}
-	settings := map[string]any{}
-	if err := json.Unmarshal(data, &settings); err != nil {
-		return fmt.Errorf("parse %s: %w", settingsPath, err)
-	}
-	stripRiffpadHooks(settings)
-	raw, err := json.MarshalIndent(settings, "", "  ")
-	if err != nil {
-		return err
-	}
-	if err := os.WriteFile(settingsPath, raw, 0o600); err != nil {
-		return err
-	}
-	fmt.Println(t.T("detach_restored"))
-	return nil
-}
-
-func defaultDaemonPort(base string) int {
-	if i := strings.LastIndex(base, ":"); i >= 0 {
-		if p, err := strconv.Atoi(base[i+1:]); err == nil {
-			return p
-		}
-	}
-	return 8787
-}
+func setupCmd(args []string, dataDir string) error { return commands.SetupCmd(args, dataDir) }
 
 // loginCmd logs into the Riffpad relay and stores the token in the daemon
 // config. `riffpad relay login` is kept as an alias.
@@ -1190,281 +665,13 @@ func openBrowser(url string) {
 // restartDaemon applies a new relay login to the running daemon. systemd-managed
 // daemons are restarted via systemctl; plain `riffpad daemon start` processes
 // are stopped and started again. A daemon that is not running is left alone.
-func restartDaemon(dataDir string) {
-	base := os.Getenv("RIFFPAD_URL")
-	if base == "" {
-		base = "http://127.0.0.1:8787"
-	}
-	if runtime.GOOS == "linux" {
-		if out, err := exec.Command("systemctl", "--user", "is-active", "riffpad.service").CombinedOutput(); err == nil && strings.TrimSpace(string(out)) == "active" {
-			if err := exec.Command("systemctl", "--user", "restart", "riffpad.service").Run(); err != nil {
-				fmt.Println(t.T("login_restart_failed", err))
-			} else {
-				fmt.Println(t.T("login_restarted"))
-			}
-			return
-		}
-	}
-	if !reachable(base) {
-		return
-	}
-	if err := daemonStop(base, dataDir); err != nil {
-		fmt.Println(t.T("login_restart_failed", err))
-		return
-	}
-	if err := daemonStart(base, dataDir); err != nil {
-		fmt.Println(t.T("login_restart_failed", err))
-		return
-	}
-	fmt.Println(t.T("login_restarted"))
-}
+// restartDaemon applies a new relay login to the running daemon; see
+// internal/commands (daemon lifecycle). Kept here only until the auth
+// migration moves its callers.
+func restartDaemon(dataDir string) { commands.RestartDaemonAfterLogin(dataDir) }
 
-func reachable(base string) bool {
-	client := &http.Client{Timeout: 500 * time.Millisecond}
-	resp, err := daemonDo(client, http.MethodGet, base+"/api/status", nil)
-	if err != nil {
-		return false
-	}
-	resp.Body.Close()
-	return resp.StatusCode == http.StatusOK
-}
+func reachable(base string) bool { return cliutil.Reachable(base) }
 
-// updateCmd checks the latest GitHub release, downloads the binary for the
-// current platform, verifies its SHA256, and atomically replaces this
-// executable (keeping a .riffpad.bak backup).
-// killCmd triggers the daemon kill switch: stops all agent sessions and
-// revokes all paired devices (local + cloud).
-func killCmd(base string) error {
-	resp, err := daemonDo(nil, http.MethodPost, base+"/api/killswitch", nil)
-	if err != nil {
-		return fmt.Errorf("daemon not reachable at %s: %w", base, err)
-	}
-	defer resp.Body.Close()
-	var out struct {
-		Killed   bool `json:"killed"`
-		Sessions int  `json:"sessions"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&out); err != nil {
-		return err
-	}
-	fmt.Println(t.T("kill_done", out.Sessions))
-	return nil
-}
+func killCmd(base string) error { return commands.KillCmd(base) }
 
-func updateCmd(args []string, dataDir string) error {
-	fs := flag.NewFlagSet("update", flag.ExitOnError)
-	force := fs.Bool("force", false, "reinstall even if already up to date")
-	noRestart := fs.Bool("no-restart", false, "do not restart the daemon after updating")
-	_ = fs.Parse(args)
-
-	fmt.Println(t.T("update_current", version.Version))
-	latest, err := latestReleaseTag()
-	if err != nil {
-		return err
-	}
-	fmt.Println(t.T("update_latest", latest))
-	if !*force && compareVersions(version.Version, latest) >= 0 {
-		fmt.Println(t.T("update_up_to_date"))
-		return nil
-	}
-
-	osName, arch, err := updatePlatform()
-	if err != nil {
-		return err
-	}
-	asset := "riffpad-" + osName + "-" + arch
-	base := updateDownloadBase
-	fmt.Println(t.T("update_downloading", asset))
-
-	exe, err := osExecutableFn()
-	if err != nil {
-		return err
-	}
-	dir := filepath.Dir(exe)
-	tmp, err := os.CreateTemp(dir, ".riffpad-update-*")
-	if err != nil {
-		return err
-	}
-	tmpPath := tmp.Name()
-	defer os.Remove(tmpPath)
-	if err := downloadFile(base+asset, tmp); err != nil {
-		return err
-	}
-	if err := tmp.Close(); err != nil {
-		return err
-	}
-	if err := verifyChecksum(base+"sha256sums.txt", asset, tmpPath); err != nil {
-		fmt.Println(t.T("update_checksum_failed"))
-		return err
-	}
-	if err := os.Chmod(tmpPath, 0o755); err != nil {
-		return err
-	}
-
-	backup := exe + ".riffpad.bak"
-	if err := copyFile(exe, backup); err != nil {
-		fmt.Println(t.T("update_backup_failed", err))
-		return err
-	}
-	if err := os.Rename(tmpPath, exe); err != nil {
-		fmt.Println(t.T("update_replace_failed", err))
-		return err
-	}
-	fmt.Println(t.T("update_done", latest, backup))
-	if !*noRestart && reachable(defaultDaemonBase()) {
-		return daemonRestart(defaultDaemonBase(), dataDir)
-	}
-	fmt.Println(t.T("update_restart_hint"))
-	return nil
-}
-
-func defaultDaemonBase() string {
-	if b := os.Getenv("RIFFPAD_URL"); b != "" {
-		return b
-	}
-	return "http://127.0.0.1:8787"
-}
-
-func latestReleaseTag() (string, error) {
-	// http.DefaultClient has no timeout; a wedged network must fail the
-	// update instead of hanging it forever (#174).
-	client := &http.Client{Timeout: 30 * time.Second}
-	resp, err := client.Get(updateAPIBase + "/" + updateRepo + "/releases/latest")
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return "", fmt.Errorf("%s", t.T("latest_release_failed", resp.Status))
-	}
-	var rel struct {
-		TagName string `json:"tag_name"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&rel); err != nil {
-		return "", err
-	}
-	if rel.TagName == "" {
-		return "", fmt.Errorf("%s", t.T("release_no_tag"))
-	}
-	return rel.TagName, nil
-}
-
-func updatePlatform() (string, string, error) {
-	osName := strings.ToLower(runtime.GOOS)
-	if osName != "linux" && osName != "darwin" {
-		return "", "", fmt.Errorf("%s", t.T("update_platform_unsupported", runtime.GOOS))
-	}
-	arch := runtime.GOARCH
-	if arch == "x86_64" {
-		arch = "amd64"
-	}
-	if arch == "aarch64" {
-		arch = "arm64"
-	}
-	if arch != "amd64" && arch != "arm64" {
-		return "", "", fmt.Errorf("%s", t.T("update_arch_unsupported", runtime.GOARCH))
-	}
-	return osName, arch, nil
-}
-
-func downloadFile(url string, w io.Writer) error {
-	// Binary downloads get a generous but bounded timeout (#174).
-	client := &http.Client{Timeout: 5 * time.Minute}
-	resp, err := client.Get(url)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("%s", t.T("download_failed", url, resp.Status))
-	}
-	_, err = io.Copy(w, resp.Body)
-	return err
-}
-
-func verifyChecksum(sumsURL, asset, path string) error {
-	client := &http.Client{Timeout: 30 * time.Second}
-	resp, err := client.Get(sumsURL)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("%s", t.T("download_failed", sumsURL, resp.Status))
-	}
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return err
-	}
-	want := ""
-	for _, line := range strings.Split(string(body), "\n") {
-		fields := strings.Fields(line)
-		if len(fields) == 2 && fields[1] == asset {
-			want = fields[0]
-			break
-		}
-	}
-	if want == "" {
-		return fmt.Errorf("%s", t.T("checksum_missing", asset))
-	}
-	data, err := os.ReadFile(path)
-	if err != nil {
-		return err
-	}
-	sum := sha256.Sum256(data)
-	got := hex.EncodeToString(sum[:])
-	if got != want {
-		return fmt.Errorf("SHA256 不匹配（期望 %s，实际 %s）", want, got)
-	}
-	return nil
-}
-
-func copyFile(src, dst string) error {
-	in, err := os.Open(src)
-	if err != nil {
-		return err
-	}
-	defer in.Close()
-	info, err := in.Stat()
-	if err != nil {
-		return err
-	}
-	out, err := os.OpenFile(dst, os.O_CREATE|os.O_WRONLY|os.O_TRUNC, info.Mode())
-	if err != nil {
-		return err
-	}
-	defer out.Close()
-	_, err = io.Copy(out, in)
-	return err
-}
-
-// compareVersions compares semver-ish versions, ignoring "v" prefixes and
-// prerelease/build suffixes (e.g. 0.1.0-m0 == v0.1.0).
-func compareVersions(a, b string) int {
-	a = strings.TrimPrefix(a, "v")
-	b = strings.TrimPrefix(b, "v")
-	if i := strings.IndexAny(a, "-+"); i >= 0 {
-		a = a[:i]
-	}
-	if i := strings.IndexAny(b, "-+"); i >= 0 {
-		b = b[:i]
-	}
-	as := strings.Split(a, ".")
-	bs := strings.Split(b, ".")
-	for i := 0; i < 3; i++ {
-		av, bv := 0, 0
-		if i < len(as) {
-			av, _ = strconv.Atoi(as[i])
-		}
-		if i < len(bs) {
-			bv, _ = strconv.Atoi(bs[i])
-		}
-		if av < bv {
-			return -1
-		}
-		if av > bv {
-			return 1
-		}
-	}
-	return 0
-}
+func updateCmd(args []string, dataDir string) error { return commands.UpdateCmd(args, dataDir) }
