+++ apps/daemon/cmd/riffpad/main_test.go
@@ -1,38 +1,14 @@
 package main
 
 import (
-	"encoding/json"
 	"net/http"
 	"net/http/httptest"
-	"os"
-	"path/filepath"
-	"strings"
-	"sync"
 	"sync/atomic"
 	"testing"
-	"time"
 
 	"github.com/riffpad/riffpad/apps/daemon/internal/cliutil"
 	"github.com/riffpad/riffpad/apps/daemon/internal/config"
-	"github.com/riffpad/riffpad/apps/daemon/internal/daemon"
 )
-
-func fakeDaemon(t *testing.T, ready *atomic.Bool) *httptest.Server {
-	t.Helper()
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/api/status" {
-			http.NotFound(w, r)
-			return
-		}
-		if ready.Load() {
-			w.WriteHeader(http.StatusOK)
-			return
-		}
-		w.WriteHeader(http.StatusServiceUnavailable)
-	}))
-	t.Cleanup(srv.Close)
-	return srv
-}
 
 func TestOAuthDeviceLoginRetriesTransientError(t *testing.T) {
 	oldOpen, oldRestart := openBrowserFn, restartDaemonFn
@@ -103,76 +79,6 @@ func TestSyncHostCredsSameUserKeepsHost(t *testing.T) {
 	}
 }
 
-func TestEnsureDaemonSkipsWhenReachable(t *testing.T) {
-	ready := &atomic.Bool{}
-	ready.Store(true)
-	srv := fakeDaemon(t, ready)
-	calls := 0
-	startDaemonFn = func(string, string) error { calls++; return nil }
-	defer func() { startDaemonFn = daemonStart }()
-
-	if err := ensureDaemon(srv.URL, t.TempDir()); err != nil {
-		t.Fatal(err)
-	}
-	if calls != 0 {
-		t.Fatalf("expected no daemon start when reachable, got %d", calls)
-	}
-}
-
-func TestEnsureDaemonStartsLazily(t *testing.T) {
-	ready := &atomic.Bool{}
-	srv := fakeDaemon(t, ready)
-	calls := 0
-	startDaemonFn = func(base, dataDir string) error {
-		calls++
-		if base != srv.URL {
-			t.Fatalf("unexpected base %q", base)
-		}
-		ready.Store(true)
-		return nil
-	}
-	defer func() { startDaemonFn = daemonStart }()
-
-	if err := ensureDaemon(srv.URL, t.TempDir()); err != nil {
-		t.Fatal(err)
-	}
-	if calls != 1 {
-		t.Fatalf("expected one daemon start, got %d", calls)
-	}
-}
-
-func TestEnsureDaemonConcurrentSingleStart(t *testing.T) {
-	ready := &atomic.Bool{}
-	srv := fakeDaemon(t, ready)
-	var mu sync.Mutex
-	calls := 0
-	startDaemonFn = func(string, string) error {
-		mu.Lock()
-		calls++
-		mu.Unlock()
-		time.Sleep(150 * time.Millisecond) // simulate slow startup
-		ready.Store(true)
-		return nil
-	}
-	defer func() { startDaemonFn = daemonStart }()
-
-	dataDir := t.TempDir()
-	var wg sync.WaitGroup
-	for i := 0; i < 4; i++ {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			if err := ensureDaemon(srv.URL, dataDir); err != nil {
-				t.Errorf("ensureDaemon: %v", err)
-			}
-		}()
-	}
-	wg.Wait()
-	if calls != 1 {
-		t.Fatalf("expected exactly one daemon start under concurrency, got %d", calls)
-	}
-}
-
 // withCliToken stubs the CLI's local token for the duration of a test.
 func withCliToken(t *testing.T, token string) {
 	t.Helper()
@@ -180,313 +86,5 @@ func withCliToken(t *testing.T, token string) {
 	t.Cleanup(func() { cliutil.SetToken("") })
 }
 
-func TestDaemonDoAttachesLocalToken(t *testing.T) {
-	withCliToken(t, "test-token")
-	var got string
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		got = r.Header.Get(daemon.LocalTokenHeader)
-		w.WriteHeader(http.StatusOK)
-	}))
-	t.Cleanup(srv.Close)
-
-	resp, err := daemonDo(nil, http.MethodGet, srv.URL+"/api/status", nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	resp.Body.Close()
-	if got != "test-token" {
-		t.Fatalf("expected %s header %q, got %q", daemon.LocalTokenHeader, "test-token", got)
-	}
-}
-
 // The CLI health check must authenticate too: a daemon that requires the
 // local token is "reachable" only when the CLI presents it.
-func TestReachableWithTokenRequiringDaemon(t *testing.T) {
-	withCliToken(t, "test-token")
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.Header.Get(daemon.LocalTokenHeader) != "test-token" {
-			w.WriteHeader(http.StatusUnauthorized)
-			return
-		}
-		w.WriteHeader(http.StatusOK)
-	}))
-	t.Cleanup(srv.Close)
-	if !reachable(srv.URL) {
-		t.Fatal("reachable should succeed with the local token")
-	}
-
-	withCliToken(t, "wrong-token")
-	if reachable(srv.URL) {
-		t.Fatal("reachable should fail with the wrong token")
-	}
-}
-
-// --- attach/detach hook merging (issue #168) ---
-
-// setupAttachEnv points HOME at a temp dir and stubs the local token so
-// attachCmd writes to a throwaway ~/.claude/settings.json.
-func setupAttachEnv(t *testing.T) (home string, base string) {
-	t.Helper()
-	home = t.TempDir()
-	t.Setenv("HOME", home)
-	withCliToken(t, "")
-	ready := &atomic.Bool{}
-	ready.Store(true)
-	return home, fakeDaemon(t, ready).URL
-}
-
-func userHookEntry(command string) map[string]any {
-	return map[string]any{
-		"matcher": "Bash",
-		"hooks":   []any{map[string]any{"type": "command", "command": command}},
-	}
-}
-
-func riffpadHookEntry(path string) map[string]any {
-	return map[string]any{
-		"matcher": "",
-		"hooks": []any{map[string]any{
-			"type": "http",
-			"url":  "http://127.0.0.1:8787/hooks/claude/" + path,
-		}},
-	}
-}
-
-func writeClaudeSettings(t *testing.T, home string, settings map[string]any) {
-	t.Helper()
-	dir := filepath.Join(home, ".claude")
-	if err := os.MkdirAll(dir, 0o700); err != nil {
-		t.Fatal(err)
-	}
-	raw, err := json.MarshalIndent(settings, "", "  ")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(dir, "settings.json"), raw, 0o600); err != nil {
-		t.Fatal(err)
-	}
-}
-
-func readClaudeSettings(t *testing.T, home string) map[string]any {
-	t.Helper()
-	raw, err := os.ReadFile(filepath.Join(home, ".claude", "settings.json"))
-	if err != nil {
-		t.Fatal(err)
-	}
-	settings := map[string]any{}
-	if err := json.Unmarshal(raw, &settings); err != nil {
-		t.Fatal(err)
-	}
-	return settings
-}
-
-// countHooks classifies every entry under settings["hooks"] as riffpad-owned
-// (URL path under /hooks/claude/) or user-owned.
-func countHooks(t *testing.T, settings map[string]any) (riffpad, user int) {
-	t.Helper()
-	hooks, _ := settings["hooks"].(map[string]any)
-	for _, list := range hooks {
-		entries, _ := list.([]any)
-		for _, e := range entries {
-			entry, _ := e.(map[string]any)
-			nested, _ := entry["hooks"].([]any)
-			isRiffpad := false
-			for _, h := range nested {
-				hm, _ := h.(map[string]any)
-				if raw, _ := hm["url"].(string); strings.Contains(raw, "/hooks/claude/") {
-					isRiffpad = true
-				}
-			}
-			if isRiffpad {
-				riffpad++
-			} else {
-				user++
-			}
-		}
-	}
-	return riffpad, user
-}
-
-func TestAttachPreservesUserHooks(t *testing.T) {
-	home, base := setupAttachEnv(t)
-	writeClaudeSettings(t, home, map[string]any{
-		"model": "opus",
-		"hooks": map[string]any{
-			"PreToolUse":   []any{userHookEntry("my-formatter")},
-			"Notification": []any{userHookEntry("my-notifier")},
-		},
-	})
-
-	if err := attachCmd(base); err == nil || !strings.Contains(err.Error(), "deprecated") {
-		t.Fatalf("expected deprecated error, got %v", err)
-	}
-	settings := readClaudeSettings(t, home)
-	if settings["model"] != "opus" {
-		t.Fatalf("unrelated setting lost: %+v", settings)
-	}
-	riffpad, user := countHooks(t, settings)
-	if riffpad != 0 || user != 2 {
-		t.Fatalf("disabled attach must not touch settings: riffpad=%d user=%d", riffpad, user)
-	}
-}
-
-func TestDetachRemovesOnlyRiffpadHooks(t *testing.T) {
-	home, _ := setupAttachEnv(t)
-	legacyHooks := map[string]any{
-		"MessageDisplay":    []any{riffpadHookEntry("message-display")},
-		"Notification":      []any{riffpadHookEntry("notification")},
-		"PermissionRequest": []any{riffpadHookEntry("permission")},
-		"PostToolUse":       []any{riffpadHookEntry("post-tool-use")},
-		"PreToolUse":        []any{riffpadHookEntry("pre-tool-use")},
-		"SessionEnd":        []any{riffpadHookEntry("session-end")},
-		"SessionStart":      []any{riffpadHookEntry("session-start")},
-		"UserPromptSubmit":  []any{riffpadHookEntry("user-prompt-submit")},
-	}
-	writeClaudeSettings(t, home, map[string]any{
-		"model": "opus",
-		"hooks": map[string]any{
-			"PreToolUse":   []any{userHookEntry("my-formatter"), riffpadHookEntry("pre-tool-use")},
-			"Notification": []any{userHookEntry("my-notifier")},
-			"PostToolUse":  legacyHooks["PostToolUse"],
-			"SessionStart": legacyHooks["SessionStart"],
-		},
-	})
-
-	if err := detachCmd(); err != nil {
-		t.Fatal(err)
-	}
-	settings := readClaudeSettings(t, home)
-	if settings["model"] != "opus" {
-		t.Fatalf("unrelated setting lost: %+v", settings)
-	}
-	riffpad, user := countHooks(t, settings)
-	if riffpad != 0 {
-		t.Fatalf("expected no riffpad entries after detach, got %d", riffpad)
-	}
-	if user != 2 {
-		t.Fatalf("expected both user hook entries kept, got %d", user)
-	}
-}
-
-// --- run command error propagation (issue #174 #1) ---
-
-func TestDaemonRestartPrefersSystemd(t *testing.T) {
-	oldActive, oldRestart := systemdActiveFn, systemdRestartFn
-	systemdActiveFn = func() (bool, error) { return true, nil }
-	var calls int
-	systemdRestartFn = func() error { calls++; return nil }
-	t.Cleanup(func() { systemdActiveFn, systemdRestartFn = oldActive, oldRestart })
-
-	if err := daemonRestart("http://127.0.0.1:1", t.TempDir()); err != nil {
-		t.Fatalf("restart failed: %v", err)
-	}
-	if calls != 1 {
-		t.Fatalf("expected one systemctl restart, got %d", calls)
-	}
-}
-
-func TestDaemonRestartRequiresRunningDaemon(t *testing.T) {
-	oldActive, oldRestart := systemdActiveFn, systemdRestartFn
-	systemdActiveFn = func() (bool, error) { return false, nil }
-	systemdRestartFn = func() error {
-		t.Fatal("must not call systemctl when service is inactive")
-		return nil
-	}
-	t.Cleanup(func() { systemdActiveFn, systemdRestartFn = oldActive, oldRestart })
-
-	err := daemonRestart("http://127.0.0.1:1", t.TempDir())
-	if err == nil || !strings.Contains(err.Error(), "daemon is not running") {
-		t.Fatalf("expected not-running error, got %v", err)
-	}
-}
-
-func TestRestartViaWindowsTaskStopsThenRuns(t *testing.T) {
-	oldRun := windowsTaskRunFn
-	var runs, shutdowns int
-	down := false
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		switch r.URL.Path {
-		case "/api/status":
-			if down {
-				w.WriteHeader(http.StatusServiceUnavailable)
-			} else {
-				w.WriteHeader(http.StatusOK)
-			}
-		case "/api/shutdown":
-			shutdowns++
-			down = true
-			w.WriteHeader(http.StatusOK)
-		default:
-			http.NotFound(w, r)
-		}
-	}))
-	t.Cleanup(srv.Close)
-	windowsTaskRunFn = func() error {
-		runs++
-		down = false // the scheduled task spawns the daemon
-		return nil
-	}
-	t.Cleanup(func() { windowsTaskRunFn = oldRun })
-
-	if err := restartViaWindowsTask(srv.URL, t.TempDir()); err != nil {
-		t.Fatalf("restart failed: %v", err)
-	}
-	if runs != 1 || shutdowns != 1 {
-		t.Fatalf("runs=%d shutdowns=%d, want 1/1", runs, shutdowns)
-	}
-}
-
-func TestRestartViaWindowsTaskSkipsStopWhenNotRunning(t *testing.T) {
-	oldRun := windowsTaskRunFn
-	var runs, shutdowns int
-	down := true
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		switch r.URL.Path {
-		case "/api/status":
-			if down {
-				w.WriteHeader(http.StatusServiceUnavailable)
-			} else {
-				w.WriteHeader(http.StatusOK)
-			}
-		case "/api/shutdown":
-			shutdowns++
-			w.WriteHeader(http.StatusOK)
-		default:
-			http.NotFound(w, r)
-		}
-	}))
-	t.Cleanup(srv.Close)
-	windowsTaskRunFn = func() error {
-		runs++
-		down = false
-		return nil
-	}
-	t.Cleanup(func() { windowsTaskRunFn = oldRun })
-
-	if err := restartViaWindowsTask(srv.URL, t.TempDir()); err != nil {
-		t.Fatalf("restart failed: %v", err)
-	}
-	if runs != 1 || shutdowns != 0 {
-		t.Fatalf("runs=%d shutdowns=%d, want 1/0", runs, shutdowns)
-	}
-}
-
-// --- daemon stop force-kill safety (issue #174 #9) ---
-
-func TestForceKillDaemonMissingPidFile(t *testing.T) {
-	if err := forceKillDaemon(t.TempDir()); err == nil {
-		t.Fatal("expected error when daemon.pid is missing")
-	}
-}
-
-func TestForceKillDaemonRefusesForeignProcess(t *testing.T) {
-	// pid 1 is never a riffpad daemon; a stale/recycled pid file must not get
-	// an unrelated process killed (#174).
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "daemon.pid"), []byte("1\n"), 0o600); err != nil {
-		t.Fatal(err)
-	}
-	if err := forceKillDaemon(dir); err == nil || !strings.Contains(err.Error(), "not a riffpad daemon") {
-		t.Fatalf("expected refusal to kill pid 1, got %v", err)
-	}
-}
