+++ apps/daemon/internal/commands/kill_unix.go
@@ -1,6 +1,6 @@
 //go:build !windows
 
-package main
+package commands
 
 import (
 	"fmt"
