+++ apps/daemon/internal/commands/lock_windows.go
@@ -1,6 +1,6 @@
 //go:build windows
 
-package main
+package commands
 
 import (
 	"os"
