+++ apps/daemon/internal/commands/procattr_unix.go
@@ -1,6 +1,6 @@
 //go:build !windows
 
-package main
+package commands
 
 import "syscall"
